implement delete pre-check for user

+++ databases/sql/tracker/050-users.sql
@@ -55,6 +55,22 @@ end;
 $$ language plpgsql;
 /*
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ * deleteUsersByIdPreCheck() function
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ */
+create or replace function deleteUsersByIdPreCheck(entityId uuid) returns boolean as
+$$
+begin
+    /*
+     * This is currently a placeholder.  This function should be overloaded later
+     * as record consumers are defined so that we can prevent any record here from being
+     * deleted until all related records are removed.
+     */
+    return true;
+end;
+$$ language plpgsql;
+/*
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
  * deleteUsersById() function
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
  */
@@ -63,11 +79,13 @@ $$
 declare
     count integer;
 begin
-    --Note: a user should not be deletable if it is associated with any
-    --      project, ticket, workflow, team (as owner)
-    delete from users where id = userId;
-    get diagnostics count = ROW_COUNT;
-    return count;
+    if deleteWorkflowPreCheck(userId) then
+        delete from users where id = userId;
+        get diagnostics count = ROW_COUNT;
+        return count;
+    else
+        return 0;
+    end if;
 end;
 $$ language plpgsql;
 /*
